Exception handling.

+++ oh-bugger.py
@@ -45,13 +45,19 @@ snd|  ,     |  /\ \ \__    |       \_
 """
 # Create text files containing the content until we hit an amount of free space left.
 i = 0
-while (usage.free / 1024) > 2:
-    file_name = f"oh-bugger_{i}.txt"
-    file_path = os.path.join(directory_path, file_name)
-    with open(file_path, "w") as file:
-        file.write(BUGGER)
-        usage = shutil.disk_usage(directory_path)
-    i += 1
+try:
+    while (usage.free / 1024) > 2:
+        file_name = f"oh-bugger_{i}.txt"
+        file_path = os.path.join(directory_path, file_name)
+        if not os.path.exists(file_path):
+            with open(file_path, "w") as file:
+                file.write(BUGGER)
+                usage = shutil.disk_usage(directory_path)
+        i += 1
+except KeyboardInterrupt as k:
+    print(f"Okay fine, here are the directory results:")
+except Exception as e:
+    print(f"An error occurred: {e}")
 
 # Print space details of drive.
 print(f"Drive: {directory_path}")
